User logo refactor
