+++ src/api/user/index.ts
@@ -1,11 +1,10 @@
 import express, { NextFunction, Request, Response } from 'express';
 
-import { UserCrud } from './user.crud';
-import { UpdateUserLogoSchema } from './validation.schema';
-
 import { ErrorMessages } from '~/core/dictionary/error.messages';
 import { UnauthorizedError, UnprocessableEntityError } from '~/core/errors';
 import { isAuthenticated } from '~/shared/user';
+import { UpdateUserLogoSchema } from '~/shared/user/schema';
+import { UserCrudService } from '~/shared/user/User.crud';
 
 const route = express.Router();
 
@@ -211,17 +210,21 @@ route.post('/logo', async (req: Request, res: Response, next: NextFunction) => {
       throw new UnprocessableEntityError(parsedBody.error.errors[0].message);
     }
 
-    await UserCrud.update({
-      logo: parsedBody.data.logo,
-      userId: user.id,
-    });
+    await UserCrudService.update(
+      {
+        logo: parsedBody.data.logo,
+      },
+      user.id,
+    );
 
     return res.status(201).json({
       type: 'USER_LOGO_UPDATED',
       statusCode: 201,
       message: `User logo updated successfully.`,
       isSuccess: true,
-      details: {},
+      details: {
+        user,
+      },
     });
   } catch (error: unknown) {
     return next(error);
@@ -300,9 +303,7 @@ route.delete(
     }
 
     try {
-      await UserCrud.delete({
-        userId: user.id,
-      });
+      await UserCrudService.deleteLogo(user.id);
 
       return res.status(200).json({
         type: 'USER_LOGO_DELETED',
